CAPSTONE_PROJECT/requirements.txt
# PostgreSQL-related packages
psycopg2-binary==2.9.6

# Apache Airflow and related packages
apache-airflow==2.3.0
apache-airflow-providers-postgres==3.0.0
apache-airflow-providers-google==7.0.0
python-dotenv==0.21.0

# Google Cloud related packages
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.0
google-auth==2.22.0

# dbt and BigQuery-related packages
dbt-bigquery==1.8.2

# Data processing and utility packages
pandas==2.1.1
sqlalchemy==2.0.20
pyyaml==6.0

# Logging and monitoring
loguru==0.7.1

# Docker
docker==6.1.3

# Additional utilities
requests==2.31.0
